+++ moor_ffi/CHANGELOG.md
@@ -1,3 +1,8 @@
+## unreleased
+
+- Remove the `background` flag from the moor apis provided by this package. Use the moor isolate api
+  instead.
+
 ## 0.0.1
 
 - Initial release. Contains standalone bindings and a moor implementation.